feature. update model
